% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{acres_to_square_meters}
\alias{acres_to_square_meters}
\title{Acres to Square Meters}
\usage{
acres_to_square_meters(acres)
}
\arguments{
\item{acres}{area in square meters}
}
\description{
Function converts area in acres to area in square meters
}
