% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{american_river_instream}
\alias{american_river_instream}
\title{American River Instream Flow to Habitat Area Relationship}
\format{dataframe with 30 rows and 4 variables:
\describe{
\item{flow_cfs}{integer flow value in cubic feet per second}
\item{FR_spawning}{Fall Run spawning WUA in square feet per 1000 feet}
\item{ST_spawning}{Steelhead spawning WUA in square feet per 1000 feet}
\item{watershed}{name of watershed}
}}
\source{
\href{https://s3-us-west-2.amazonaws.com/cvpiahabitat-r-package/cvpia-sit-model-inputs/FWS2003_AmericanRiverPHABSIM2DFinalReport.pdf}{FWS 2003}
}
\usage{
american_river_instream
}
\description{
A dataset containing the Weighted Usable Area (WUA) in square feet per 1000 feet
as a function of flow in cubic feet per second
}
\details{
The spawning WUA was derived from a River2D,
applied habitat suitability model.
}
\keyword{datasets}
