% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-floodplain-habitat.R
\name{apply_suitability}
\alias{apply_suitability}
\title{Apply Suitability Factor}
\usage{
apply_suitability(fp_hab_sq_meters, suitable_factor = 0.27)
}
\arguments{
\item{fp_hab_sq_meters}{Total floodplain area in square meters}

\item{suitable_factor}{Habitat suitability factor, default value of .27 from
U.S. Fish and Wildlife Services San Joaquin River Restoration Program
\href{https://s3-us-west-2.amazonaws.com/cvpiahabitat-r-package/20121127_MinimumRearingHabitat.pdf}{2012 report.}}
}
\value{
Total suitable floodplain area in square meters
}
\description{
The modeled flow to floodplain area relationships within the package are for total innundated area.
This function applys a suitability factor to the total area in order to estimate total suitable floodplain area
available at a given flow.
}
