% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{battle_creek_floodplain}
\alias{battle_creek_floodplain}
\title{Battle Creek Flow to Floodplain Habitat Area Relationship}
\format{dataframe with 22 rows and 3 variables:
\describe{
  \item{flow_cfs}{integer flow value in cubic feet per second}
  \item{floodplain_acres}{fall run floodplain acreage (use for Spring Run and Steelhead when required)}
  \item{watershed}{name of watershed}
}}
\source{
\href{https://s3-us-west-2.amazonaws.com/cvpiahabitat-r-package/cvpia-sit-model-inputs/DRAFT+Deer+Creek+Hydraulic+Models+Tech+Memo+6-08-07.pdf}{Scaled from a Deer Creek flow to floodplain area relationship generated with a 1D HEC-RAS hydraulic model}
}
\usage{
battle_creek_floodplain
}
\description{
A dataset containing total inundated floodplain area in acres as a function of flow in cubic feet per second.
}
\details{
HEC-RAS model runs to generate Deer Creek CVPIA DSM inputs conducted by Mark Gard (mark_gard@fws.gov) in 2017. A Battle Creek scaling factor of 147\% was calculated as the average ratio of Battle Creek monthly mean flow to Deer Creek monthly mean flow between December and June. This scaling factor was applied to Deer Creek flows and corresponding floodplain area per river mile values, which were then multiplied by the 5.87 mile low gradient mapped rearing extent in Battle Creek. Remote sensing analysis of aerial photography was used to confirm differentiation between instream (active channel area of 52.4 acres) and floodplain inundation area.
}
\examples{
battle_creek_floodplain

}
\keyword{datasets}
