% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{battle_creek_instream}
\alias{battle_creek_instream}
\title{Battle Creek Instream Flow to Habitat Area Relationship}
\format{dataframe with 35 rows and 6 variables:
\describe{
  \item{flow_cfs}{integer flow value in cubic feet per second}
  \item{spawn_WUA}{spawning WUA in square feet per 1000 feet}
  \item{fry_WUA}{fry (up to 50 mm) WUA in square feet per 1000 feet}
  \item{juv_WUA}{juvenile WUA in square feet per 1000 feet}
  \item{adult_trout_WUA}{adult trout WUA in square feet per 1000 feet}
  \item{watershed}{name of watershed}
}}
\source{
\href{https://s3-us-west-2.amazonaws.com/cvpiahabitat-r-package/cvpia-sit-model-inputs/Payne1995_BattleCreekIFIM.pdf}{Payne 1995}
}
\usage{
battle_creek_instream
}
\description{
A dataset containing the Weighted Usable Area (WUA) in square feet per 1000 feet
as a function of flow in cubic feet per second
}
\details{
The spawning and rearing WUA values were derived from a PHABSIM model. The results are for Fall Run Chinook.
}
\keyword{datasets}
