% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clear_creek_floodplain}
\alias{clear_creek_floodplain}
\title{Clear Creek Flow to Floodplain Habitat Area Relationship}
\format{dataframe with 8 rows and 3 variables:
\describe{
  \item{flow_cfs}{integer flow value in cubic feet per second}
  \item{floodplain_acres}{fall run floodplain acreage (use for Spring Run and Steelhead when required)}
  \item{watershed}{name of watershed}
}}
\source{
NA
}
\usage{
clear_creek_floodplain
}
\description{
A dataset containing total inundated floodplain area in acres as a function of flow in cubic feet per second
}
\details{
HEC-RAS model runs to generate Cottonwood Creek CVPIA DSM inputs conducted by Mark Gard (mark_gard@fws.gov) in 2017. A Clear Creek scaling factor of 21\% was calculated as the average ratio of Clear Creek monthly mean flow to Cottonwood Creek monthly mean flow between December and June. This scaling factor was applied to Cottonwood Creek flows and corresponding floodplain area per river mile values, which were then multiplied by the 8 mile mapped rearing extent in Clear Creek. Remote sensing analysis of aerial photography was used to confirm differentiation between instream (active channel area of 57.4 acres) and floodplain inundation area.
}
\examples{
clear_creek_floodplain

}
\keyword{datasets}
