% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cottonwood_creek_floodplain}
\alias{cottonwood_creek_floodplain}
\title{Cottonwood Creek Flow to Floodplain Habitat Area Relationship}
\format{dataframe with 32 rows and 3 variables:
\describe{
  \item{flow_cfs}{integer flow value in cubic feet per second}
  \item{FR_floodplain_acres}{fall run floodplain acreage}
  \item{watershed}{name of watershed}
}}
\source{
FWS \url{https://s3-us-west-2.amazonaws.com/cvpiahabitat-r-package/cvpia-sit-model-inputs/FWS2014_CVPIA_Annual_Progress_Report_Fiscal_Year_2013.pdf}
}
\usage{
cottonwood_creek_floodplain
}
\description{
A dataset containing the floodplain habitat area in acres as a
function of flow in cubic feet per second
}
\details{
The habitat area was derived from the FWS 2017 HEC-RAS 1D model.
}
\keyword{datasets}
