% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cottonwood_creek_instream}
\alias{cottonwood_creek_instream}
\title{Cottonwood Creek Instream Flow to Habitat Area Relationship}
\format{dataframe with 36 rows and 5 variables:
\describe{
  \item{flow_cfs}{integer flow value in cubic feet per second}
  \item{spawn_WUA}{spawning WUA in square feet per 1000 feet}
  \item{fry_WUA}{fry (up to 50 mm) WUA in square feet per 1000 feet}
  \item{juv_WUA}{juvenile WUA in square feet per 1000 feet}
  \item{watershed}{name of watershed}
}}
\source{
\href{https://s3-us-west-2.amazonaws.com/cvpiahabitat-r-package/cvpia-sit-model-inputs/CDFW1979_CottonwoodSpawningIFIM.PDF}{CDFW 1979 and FWS 2014}
}
\usage{
cottonwood_creek_instream
}
\description{
A dataset containing the Weighted Usable Area (WUA) in square feet per 1000 feet
as a function of flow in cubic feet per second
}
\details{
The spawning WUA values were derived from a PHABSIM model. The rearing values are from the FWS 2014
River2D model. The results are for Fall Run.
}
\keyword{datasets}
