% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cow_creek_instream}
\alias{cow_creek_instream}
\title{Cow Creek Instream Flow to Habitat Area Relationship}
\format{dataframe with 30 rows and 4 variables:
\describe{
  \item{flow_cfs}{integer flow value in cubic feet per second}
  \item{FR_fry_wua}{fall run fry (up to 50 mm) WUA in square feet per 1000 feet}
  \item{FR_juv_wua}{fall run juvenile WUA in square feet per 1000 feet}
  \item{watershed}{name of watershed}
}}
\source{
\href{https://s3-us-west-2.amazonaws.com/cvpiahabitat-r-package/cvpia-sit-model-inputs/FWS2011_SouthCowrpt.pdf}{FWS 2011}
}
\usage{
cow_creek_instream
}
\description{
A dataset containing the Weighted Usable Area (WUA) in square feet per 1000 feet
as a function of flow in cubic feet per second
}
\details{
The rearing WUA values are from a River2D model done on South Cow Creek. The values are from two river
segments, Valley Floor Reach (5.11 miles) and Boero Reach (1.68 miles).
}
\examples{
cow_creek_instream

}
\keyword{datasets}
