% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deer_creek_floodplain}
\alias{deer_creek_floodplain}
\title{Deer Creek Flow to Floodplain Habitat Area Relationship}
\format{dataframe with 30 rows and 3 variables:
\describe{
  \item{flow_cfs}{integer flow value in cubic feet per second}
  \item{FR_floodplain_acres}{fall run floodplain acreage}
  \item{watershed}{name of watershed}
}}
\source{
TODO
}
\usage{
deer_creek_floodplain
}
\description{
A dataset containing the floodplain habitat area in acres as a
function of flow in cubic feet per second
}
\details{
The habitat area was derived from the Deer Creek Watershed Conservancy HEC-RAS 1D 2011 model.
}
\keyword{datasets}
