% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{delta_habitat}
\alias{delta_habitat}
\title{Delta Habitat Area}
\format{
dataframe with 372 rows and 3 variables
\describe{
  \item{date}{modeled results for 1980-2010}
  \item{North Delta}{high quality habitat area in square meters}
  \item{South Delta}{high quality habitat area in square meters}
}
}
\source{
\href{http://cvpia-habitat-docs-markdown.s3-website-us-west-2.amazonaws.com/watershed/delta.html}{North and South Delta Modeling Details}

\itemize{
  \item Modeling Output: Correigh Green \email{correigh.greene@noaa.gov}
  \item Delta Node Selection: Mark Tompkins \email{mtompkins@flowwest.com}
  \item Data Wrangling: Sadie Gill \email{sgill@flowwest.com}
}
}
\usage{
delta_habitat
}
\description{
A dataset containing the area of highly suitable habitat within the
North and South Deltas
}
\keyword{datasets}
