% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{feather_river_floodplain}
\alias{feather_river_floodplain}
\title{Feather River Flow to Floodplain Habitat Area Relationship}
\format{
dataframe with 40 rows and 5 variables:
\describe{
  \item{flow_cfs}{integer flow value in cubic feet per second}
  \item{FR_floodplain_acres}{fall run floodplain acreage}
  \item{SR_floodplain_acres}{spring run floodplain acreage}
  \item{ST_floodplain_acres}{steelhead floodplain acreage}
  \item{watershed}{name of watershed}
}
}
\source{
\href{https://s3-us-west-2.amazonaws.com/cvpiahabitat-r-package/cvpia-sit-model-inputs/CombinedTM_IQAR_Final-FULL-REPORT_20140206.pdf}{Central Valley Floodplain Evaluation and Delineation (CVFED) HEC-RAS hydraulic model}
}
\usage{
feather_river_floodplain
}
\description{
A dataset containing the floodplain habitat area in acres as a
function of flow in cubic feet per second
}
\section{Fall Run, Spring Run and Steelhead Modeling}{

The entire mapped rearing extent of 64.98 miles was modeled using Central Valley Floodplain Evaluation and Delineation (CVFED) HEC-RAS hydraulic model. Active channel area of 2363.20 acres estimated through remote sensing analysis was subtracted from total inundated area to get inundated floodplain area.
}

\examples{
feather_river_floodplain

}
\keyword{datasets}
