% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-instream-habitat.R, R/set-instream.R
\name{FR_rearing_approx}
\alias{FR_rearing_approx}
\alias{FR_rearing_approx}
\title{Fall Run rearing habitat flow to area approximator}
\usage{
FR_rearing_approx(relationship_df, modeling_lookup, life_stage)

FR_rearing_approx(relationship_df, modeling_lookup, life_stage)
}
\arguments{
\item{relationship_df}{dataframe from cvpiaHabitat with a flow to wua relationship}

\item{modeling_lookup}{modeling lookup dataframe from cvpiaHabitat}

\item{life_stage}{one of 'spawn', 'juv' or 'fry'}

\item{relationship_df}{dataframe from cvpiaHabitat with a flow to wua relationship}

\item{modeling_lookup}{modeling lookup dataframe from cvpiaHabitat}

\item{life_stage}{one of 'spawn', 'juv' or 'fry'}
}
\description{
function creates the approx function for fall run

function creates the approx function for fall run
}
