% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_wua_selector}
\alias{get_wua_selector}
\title{Get WUA Selector}
\usage{
get_wua_selector(species_wuas, species, life_stage)
}
\arguments{
\item{species_wuas}{vector of column names within modeling data table for the targeted watershed}

\item{species}{target species: "fr" for fall run, "sr" for spring run, and "st" for steelhead}

\item{life_stage}{"spawn", fry", "juv", or "adult"}
}
\value{
column name of desired habitat relationship
}
\description{
Habitat Modeling Column Lookup
}
\details{
Habitat modeling column lookup by species and lifestage. If desired
combination of species and lifestage is not present in modeling data table for a watershed,
then the column representing the most appropriate proxy will be returned.
}
