% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{habitat_data}
\alias{habitat_data}
\alias{american_river_instream}
\alias{american_river_floodplain}
\alias{antelope_creek_floodplain}
\alias{battle_creek_instream}
\alias{battle_creek_floodplain}
\alias{bear_creek_floodplain}
\alias{bear_river_instream}
\alias{bear_river_floodplain}
\alias{big_chico_creek_floodplain}
\alias{butte_creek_instream}
\alias{butte_creek_floodplain}
\alias{calaveras_river_instream}
\alias{calaveras_river_floodplain}
\alias{clear_creek_instream}
\alias{clear_creek_floodplain}
\alias{cosumnes_river_instream}
\alias{cosumnes_river_floodplain}
\alias{cottonwood_creek_instream}
\alias{cottonwood_creek_floodplain}
\alias{cow_creek_instream}
\alias{cow_creek_floodplain}
\alias{deer_creek_floodplain}
\alias{elder_creek_floodplain}
\alias{feather_river_instream}
\alias{feather_river_floodplain}
\alias{merced_river_instream}
\alias{merced_river_floodplain}
\alias{mill_creek_floodplain}
\alias{mokelumne_river_instream}
\alias{mokelumne_river_floodplain}
\alias{paynes_creek_floodplain}
\alias{san_joaquin_river_instream}
\alias{san_joaquin_river_floodplain}
\alias{stanislaus_river_instream}
\alias{stanislaus_river_floodplain}
\alias{stony_creek_floodplain}
\alias{thomes_creek_floodplain}
\alias{tuolumne_river_instream}
\alias{tuolumne_river_floodplain}
\alias{yuba_river_instream}
\alias{yuba_river_floodplain}
\alias{upper_mid_sac_region_instream}
\title{Habitat Data}
\description{
Datasets used to calculate spawning, instream, and floodplain
rearing habitat areas for Chinook Salmon and Steelhead
}
\section{Instream}{

Datasets containing the Weighted Usable Area (WUA) in square feet per 1000 feet
as a function of flow in cubic feet per second for Fall Run (FR), Spring Run (SR),
Winter Run (WR), and Steelhead (ST) depending on species occurance and modeling availability.

Typical data structure for streams:
\itemize{
  \item{\strong{flow_cfs} - flow value in cubic feet per second}
  \item{\strong{FR_spawn_wua} - Spawning WUA in square feet per 1000 feet}
  \item{\strong{FR_fry_wua} - Fry (up to 50 mm) WUA in square feet per 1000 feet}
  \item{\strong{FR_juv_wua} - Juvenile WUA in square feet per 1000 feet}
  \item{\strong{watershed} - Name of watershed}
}
}

\section{Floodplain}{

Datasets containing the total inuncated area in acres as a function of flow in cubic
feet per second for Fall Run (FR), Spring Run (SR), Winter Run (WR), and
Steelhead (ST) depending on species occurance and modeling availability.

Typical data structure for streams:
\itemize{
  \item{\strong{flow_cfs} - Flow value in cubic feet per second}
  \item{\strong{FR_floodplain_acres} - Total inundated area in acres for Fall Run}
  \item{\strong{SR_floodplain_acres} - Total inundated area in acres for Spring Run}
  \item{\strong{ST_floodplain_acres} - Total inundated area in acres for Steelhead}
  \item{\strong{watershed} - Name of watershed}
}
}

\section{Modeling Details for Streams}{

\itemize{
  \item \href{http://cvpia-habitat-docs-markdown.s3-website-us-west-2.amazonaws.com/watershed/american_river.html}{American River}
  \item \href{http://cvpia-habitat-docs-markdown.s3-website-us-west-2.amazonaws.com/watershed/antelope_creek.html}{Antelope Creek}
  \item \href{http://cvpia-habitat-docs-markdown.s3-website-us-west-2.amazonaws.com/watershed/battle_creek.html}{Battle Creek}
  \item \href{http://cvpia-habitat-docs-markdown.s3-website-us-west-2.amazonaws.com/watershed/bear_creek.html}{Bear Creek}
  \item \href{http://cvpia-habitat-docs-markdown.s3-website-us-west-2.amazonaws.com/watershed/bear_river.html}{Bear River}
  \item \href{http://cvpia-habitat-docs-markdown.s3-website-us-west-2.amazonaws.com/watershed/big_chico_creek.html}{Big Chico Creek}
  \item \href{http://cvpia-habitat-docs-markdown.s3-website-us-west-2.amazonaws.com/watershed/butte_creek.html}{Butte Creek}
  \item \href{http://cvpia-habitat-docs-markdown.s3-website-us-west-2.amazonaws.com/watershed/calaveras_river.html}{Calaveras River}
  \item \href{http://cvpia-habitat-docs-markdown.s3-website-us-west-2.amazonaws.com/watershed/clear_creek.html}{Clear Creek}
  \item \href{http://cvpia-habitat-docs-markdown.s3-website-us-west-2.amazonaws.com/watershed/cosumnes_river.html}{Cosumnes River}
  \item \href{http://cvpia-habitat-docs-markdown.s3-website-us-west-2.amazonaws.com/watershed/cottonwood_creek.html}{Cottonwood Creek}
  \item \href{http://cvpia-habitat-docs-markdown.s3-website-us-west-2.amazonaws.com/watershed/cow_creek.html}{Cow Creek}
  \item \href{http://cvpia-habitat-docs-markdown.s3-website-us-west-2.amazonaws.com/watershed/deer_creek.html}{Deer Creek}
  \item \href{http://cvpia-habitat-docs-markdown.s3-website-us-west-2.amazonaws.com/watershed/elder_creek.html}{Elder Creek}
  \item \href{http://cvpia-habitat-docs-markdown.s3-website-us-west-2.amazonaws.com/watershed/feather_river.html}{Feather River}
  \item \href{http://cvpia-habitat-docs-markdown.s3-website-us-west-2.amazonaws.com/watershed/merced_river.html}{Merced River}
  \item \href{http://cvpia-habitat-docs-markdown.s3-website-us-west-2.amazonaws.com/watershed/mill_creek.html}{Mill Creek}
  \item \href{http://cvpia-habitat-docs-markdown.s3-website-us-west-2.amazonaws.com/watershed/mokelumne_river.html}{Mokelumne River}
  \item \href{http://cvpia-habitat-docs-markdown.s3-website-us-west-2.amazonaws.com/watershed/paynes_creek.html}{Paynes Creek}
  \item \link[=upper_sac_ACID_boards_in]{Sacramento River}
  \item \href{http://cvpia-habitat-docs-markdown.s3-website-us-west-2.amazonaws.com/watershed/san_joaquin_river.html}{San Joaquin River}
  \item \href{http://cvpia-habitat-docs-markdown.s3-website-us-west-2.amazonaws.com/watershed/stanislaus_river.html}{Stanislaus River}
  \item \href{http://cvpia-habitat-docs-markdown.s3-website-us-west-2.amazonaws.com/watershed/stony_creek.html}{Stony Creek}
  \item \href{http://cvpia-habitat-docs-markdown.s3-website-us-west-2.amazonaws.com/watershed/thomes_creek.html}{Thomes Creek}
  \item \href{http://cvpia-habitat-docs-markdown.s3-website-us-west-2.amazonaws.com/watershed/tuolumne_river.html}{Tuolumne River}
  \item \href{http://cvpia-habitat-docs-markdown.s3-website-us-west-2.amazonaws.com/watershed/yuba_river.html}{Yuba River}
}
}

\section{Bypasses}{

\itemize{
  \item \link[=bypass]{Sutter and Yolo Bypasses}
}
}

\section{Delta}{

\itemize{
  \item \link[=delta_habitat]{North and South Delta}
}
}

\keyword{datasets}
