% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IChab.adult}
\alias{IChab.adult}
\title{Steelhead Adult Rearing Habitat}
\format{array[watersheds, months, years]}
\source{
\itemize{
  \item Wrangling: Sadie Gill \email{sgill@flowwest.com}
  \item Method: Mark Tompkins \email{mtompkins@flowwest.com}
}
}
\usage{
IChab.adult
}
\description{
A dataset containing adult rearing habitat quanities in square meters
}
\details{
Only 5 of the watersheds have wua relationships:
**Adult Trout**
1. Battle Creek
2. Butte Creek

**Steelhead**
3. Merced River
4. Tuolumne River
5. Yuba River

These watershed's habitat values vary by month.

For the other watersheds, we created one habitat value using the calsim mean flows for the years 1980-1999.
We used Merced's adult steelhead wua curve and scaled it by the ratio of mean flow December through May
between itself and Merced.
}
\keyword{datasets}
