% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lower_mid_sacramento_river_instream}
\alias{lower_mid_sacramento_river_instream}
\title{Lower-Mid Sacramento River Instream Flow to Habitat Area Relationship}
\format{dataframe with 45 rows and 3 variables:
\describe{
  \item{flow_cfs}{integer flow value in cubic feet per second}
  \item{FR_juv_wua}{juvenile WUA in square feet per 1000 feet}
  \item{watershed}{name of watershed}
}}
\source{
NMFS Life Cycle Model, Hendrix 2017
}
\usage{
lower_mid_sacramento_river_instream
}
\description{
A dataset containing the Weighted Usable Area (WUA) in square feet per 1000 feet
as a function of flow in cubic feet per second
}
\details{
The rearing WUA was derived from a HEC-RAS 1D model. The CVPIA Lower-mid Sacramento River extends
from Wilkins Slough to the American River (58.0 mi). This reach overlaps with two of the study's, Battle Creek to
the confluence with the Feather River (189.1 mi) and the confluence with the Feather River to Freeport (33.4 mi).
To scale the study's results to CVPIA's extents, we calculate the proportion of the Lower-mid Sacramento River
above (34\%) and below (66\%) the Feather River. Using these values we calculate a weighted average WUA from both of the studies.
}
\keyword{datasets}
