% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lower_sacramento_river_instream}
\alias{lower_sacramento_river_instream}
\title{Lower Sacramento River Instream Flow to Habitat Area Relationship}
\format{dataframe with 45 rows and 3 variables:
\describe{
  \item{flow_cfs}{integer flow value in cubic feet per second}
  \item{FR_juv_wua}{juvenile WUA in square feet per 1000 feet}
  \item{watershed}{name of watershed}
}}
\source{
NMFS Life Cycle Model, Hendrix 2017
}
\usage{
lower_sacramento_river_instream
}
\description{
A dataset containing the Weighted Usable Area (WUA) in square feet per 1000 feet
as a function of flow in cubic feet per second
}
\details{
The rearing WUA was derived from a HEC-RAS 1D model. The study's extent is from
the confluence with the Feather River to Freeport (33.4 mi). The CVPIA Lower Sacramento River extends
from the confluence with the American River to Freeport (13.7 mi).
}
\keyword{datasets}
