% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{merced_river_floodplain}
\alias{merced_river_floodplain}
\title{Merced River Flow to Floodplain Habitat Area Relationship}
\format{dataframe with 23 rows and 3 variables: \describe{
  \item{flow_cfs}{integer flow value in cubic feet per second}
  \item{FR_floodplain_acres}{fall run floodplain acreage (use for Steelhead when required)}
  \item{watershed}{name of watershed} }}
\source{
\href{https://s3-us-west-2.amazonaws.com/cvpiahabitat-r-package/cvpia-sit-model-inputs/CVFPP2012_Att8_June.pdf}{CVFPP2012}
}
\usage{
merced_river_floodplain
}
\description{
A dataset containing the floodplain habitat area in acres as a
  function of flow in cubic feet per second
}
\details{
The habitat area was derived from the CVFPP 2012 HEC-RAS 1D model.
  The study only represents the lower half of Merced, scale up to
  represent the whole rearing region.
}
\keyword{datasets}
