% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{merced_river_instream}
\alias{merced_river_instream}
\title{Merced River Instream Flow to Habitat Area Relationship}
\format{dataframe with 30 rows and 6 variables:
\describe{
  \item{flow_cfs}{integer flow value in cubic feet per second}
  \item{spawn_WUA}{spawning WUA in square feet per 1000 feet}
  \item{fry_WUA}{fry (up to 50 mm) WUA in square feet per 1000 feet}
  \item{juv_WUA}{juvenile WUA in square feet per 1000 feet}
  \item{adult_steelhead_WUA}{adult steelhead WUA in square feet per 1000 feet}
  \item{watershed}{name of watershed}
}}
\source{
\href{https://s3-us-west-2.amazonaws.com/cvpiahabitat-r-package/cvpia-sit-model-inputs/FWS1996_AmericanRiverSpawningFinalReport.pdf}{FWS 1996 and MID 2013}
}
\usage{
merced_river_instream
}
\description{
A dataset containing the Weighted Usable Area (WUA) in square feet per 1000 feet
as a function of flow in cubic feet per second
}
\details{
The spawning WUA values were derived from the FWS 1996 PHABSIM model and the rearing from
the MID 2013 PHABSIM.
}
\keyword{datasets}
