% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mill_creek_floodplain}
\alias{mill_creek_floodplain}
\title{Mill Creek Flow to Floodplain Habitat Area Relationship}
\format{dataframe with 22 rows and 3 variables:
\describe{
  \item{flow_cfs}{integer flow value in cubic feet per second}
  \item{floodplain_acres}{fall run floodplain acreage (use for Spring Run and Steelhead when required)}
  \item{watershed}{name of watershed}
}}
\source{
\href{https://s3-us-west-2.amazonaws.com/cvpiahabitat-r-package/cvpia-sit-model-inputs/DRAFT+Deer+Creek+Hydraulic+Models+Tech+Memo+6-08-07.pdf}{Scaled from a Deer Creek flow to floodplain area relationship generated with a 1D HEC-RAS hydraulic model}
}
\usage{
mill_creek_floodplain
}
\description{
A dataset containing total inundated floodplain area in acres as a function of flow in cubic feet per second.
}
\details{
HEC-RAS model runs to generate Deer Creek CVPIA DSM inputs conducted by Mark Gard (mark_gard@fws.gov) in 2017. A Mill Creek scaling factor of 98% was calculated as the average ratio of Mill Creek monthly mean flow to Deer Creek monthly mean flow between December and June. This scaling factor was applied to Deer Creek flows and corresponding floodplain area per river mile values, which were then multiplied by the 7.5 mile low gradient and 13.6 miles of high gradient mapped rearing extent in Mill Creek. Remote sensing analysis of aerial photography was used to confirm differentiation between instream (active channel area of 119.9 acres) and floodplain inundation area.
}
\examples{
mill_creek_floodplain

}
\keyword{datasets}
