% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{modeling_exist}
\alias{modeling_exist}
\title{Habitat Modeling Status}
\format{
dataframe with 32 rows and 23 variables
\describe{
\item{Order}{integer value representing watershed order in SIT model}
\item{Watershed}{name of watershed}
\item{FR_spawn}{TRUE if spawning habitat modeling exists for Fall Run}
\item{FR_fry}{TRUE if fry rearing habitat modeling exists for Fall Run}
\item{FR_juv}{TRUE if juvenile rearing habitat modeling exists for Fall Run}
\item{FR_floodplain}{TRUE if floodplain rearing modeling exists for Fall Run}
\item{SR_spawn}{TRUE if spawning habitat modeling exists for Spring Run}
\item{SR_fry}{TRUE if fry rearing habitat modeling exists for Spring Run}
\item{SR_juv}{TRUE if juvenile rearing habitat modeling exists for Spring Run}
\item{SR_floodplain}{TRUE if floodplain rearing modeling exists for Spring Run}
\item{ST_spawn}{TRUE if spawning habitat modeling exists for Steelhead}
\item{ST_fry}{TRUE if fry rearing habitat modeling exists for Steelhead}
\item{ST_juv}{TRUE if juvenile rearing habitat modeling exists for Steelhead}
\item{ST_floodplain}{TRUE if floodplain rearing modeling exists for Steelhead}
\item{ST_adult}{TRUE if adult rearing habitat modeling exists for Steelhead}
\item{LFR_spawn}{TRUE if spawning habitat modeling exists for Late-fall Run}
\item{LFR_fry}{TRUE if fry rearing habitat modeling exists for Late-fall Run}
\item{LFR_juv}{TRUE if juvenile rearing habitat modeling exists for Late-fall Run}
\item{LFR_floodplain}{TRUE if floodplain rearing modeling exists for Late-fall Run}
\item{WR_spawn}{TRUE if spawning habitat modeling exists for Winter Run}
\item{WR_fry}{TRUE if fry rearing habitat modeling exists for Winter Run}
\item{WR_juv}{TRUE if juvenile rearing habitat modeling exists for Winter Run}
\item{WR_floodplain}{TRUE if floodplain rearing modeling exists for Winter Run}
}
}
\source{
Sadie Gill \email{sgill@flowwest.com}
}
\usage{
modeling_exist
}
\description{
This table summaries modeling availability based on species and
lifestage combinations for each watershed
}
\examples{
modeling_exist

}
\keyword{datasets}
