% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{north_delta_floodplain}
\alias{north_delta_floodplain}
\title{North Delta Flow to Floodplain Habitat Area Relationship}
\format{dataframe with 73 rows and 3 variables:
\describe{
  \item{flow_cfs}{integer flow value in cubic feet per second}
  \item{FR_floodplain_acres}{fall run floodplain acreage (use for Spring Run and Steelhead when required)}
  \item{watershed}{name of watershed}
}}
\source{
\href{https://s3-us-west-2.amazonaws.com/cvpiahabitat-r-package/cvpia-sit-model-inputs/HendrixEtAl2014_Winter_Run_Model_Tech_Memo.pdf}{Delta Simulation Model II (DSM2) one-dimensional hydrodynamic model refined for use in the NOAA-NMFS Winter Run Chinook Salmon life cycle model}
}
\usage{
north_delta_floodplain
}
\description{
A dataset containing the floodplain habitat area in acres as a
function of flow in cubic feet per second
}
\section{Fall Run Modeling}{

The entire mapped rearing extent of 45.65 miles was modeled using Delta Simulation Model II (DSM2) one-dimensional hydrodynamic model refined for use in the NOAA-NMFS Winter Run Chinook Salmon life cycle model. Active channel area of NA acres estimated through remote sensing analysis was subtracted from total inundated area to get inundated floodplain area.
}

\examples{
north_delta_floodplain

}
\keyword{datasets}
