% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pools}
\alias{pools}
\title{Pool Area}
\format{dataframe with 31 rows and 3 variables
\describe{
  \item{watershed}{CVPIA watershed}
  \item{SR_pools_sq_meters}{Spring Run pool area estimate in square meters}
  \item{ST_pools_sq_meters}{Steelhead pool area estimate in square meters}
}}
\source{
\itemize{
  \item Percent Pool Estimates: Mark Gard \email{mark_gard@fws.gov}
  \item Channel Area Estimates: Sadie Gill \email{sgill@flowwest.com}
  \item QA/QC: Mark Tompkins \email{mtompkins@flowwest.com}
}
}
\usage{
pools
}
\description{
A dataset containing estimated area of pools within a watershed
}
\details{
Estimates of percent pools were provided by Mark Gard for some watersheds.
The mean percent pools, excluding Feather River, is 23.1\%. This value is used for watersheds without an estimate.
The total area of pools is then caluculated
by multiplying the estimated percent pools by the total channel area. Channel
areas were estimated using Google Earth Engine.

Watersheds With Estimated Percent Pools:
\itemize{
  \item Antelope Creek
  \item Battle Creek
  \item Butte Creek
  \item Clear Creek
  \item Cottonwood Creek
  \item Deer Creek
  \item Feather River (*80% percent pools seems high, but was confirmed by Jason Kindopp)
  \item Mill Creek
  \item Mokelumne River
  \item Stanislaus River
  \item Tuolumne River
  \item Upper Sacramento River
  \item Yuba River
}

\strong{Note:} The area of pools that covers both the Upper Sacramento and
Upper-mid Sacramento DSM segments is assigned to the Upper Sacramento River.
}
\keyword{datasets}
