% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-instream-habitat.R
\name{region_rearing_approx}
\alias{region_rearing_approx}
\title{Regional rearing habitat flow to area approximator}
\usage{
region_rearing_approx(region, species, life_stage)
}
\arguments{
\item{region}{Region name, example "Upper-mid Sacramento River"}

\item{species}{one of 'fr' (Fall Run), 'sr' (Spring Run), or 'st' (Steelhead)}
}
\value{
a list of approx functions obtained from calliong spawning_approx()
}
\description{
function uses a region to return approx functions for watersheds within it with models
}
