% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-spawning-habitat.R
\name{region_spawn_approx}
\alias{region_spawn_approx}
\title{function uses a region to return approx functions for watersheds within it with models}
\usage{
region_spawn_approx(region, species)
}
\arguments{
\item{region}{Region name, example "Upper-mid Sacramento River"}

\item{species}{one of 'fr' (Fall Run), 'sr' (Spring Run), or 'st' (Steelhead)}
}
\value{
a list of approx functions obtained from calliong spawning_approx()
}
\description{
function uses a region to return approx functions for watersheds within it with models
}
