% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sacramento_habitat}
\alias{sacramento_habitat}
\alias{upper_sac_ACID_boards_in}
\alias{upper_sac_ACID_boards_out}
\alias{upper_sacramento_river_instream}
\alias{upper_sacramento_river_floodplain}
\alias{upper_mid_sacramento_river_instream}
\alias{upper_mid_sacramento_river_floodplain}
\alias{lower_mid_sacramento_river_instream}
\alias{lower_mid_sacramento_river_floodplain}
\alias{lower_sacramento_river_instream}
\alias{lower_sacramento_river_floodplain}
\title{Mainstem Sacramento River Habitat Data}
\source{
\href{http://cvpia-habitat-docs-markdown.s3-website-us-west-2.amazonaws.com/watershed/sacramento_river.html}{Sacramento River Modeling Details}
}
\description{
Datasets used to calculate spawning, instream, and floodplain
rearing habitat areas for Chinook Salmon and Steelhead
}
\section{Spawning}{

Datasets containing the Weighted Usable Area (WUA) in square feet per 1000 feet
as a function of flow in cubic feet per second for Fall Run (FR), Late-fall Run (LFR),
Winter Run (WR), and Steelhead (ST).

\itemize{
  \item \code{upper_sac_ACID_boards_in}
  \item \code{upper_sac_ACID_boards_out}
}

Data Format:
\itemize{
  \item{\strong{flow_cfs} - flow value in cubic feet per second}
  \item{\strong{FR_spawn_wua} - Spawning WUA in square feet per 1000 feet (Fall Run)}
  \item{\strong{LFR_spawn_wua} - Spawning WUA in square feet per 1000 feet (Late-Fall Run)}
  \item{\strong{ST_spawn_wua} - Spawning WUA in square feet per 1000 feet (Steelhead)}
  \item{\strong{WR_spawn_wua} - Spawning WUA in square feet per 1000 feet (Winter Run)}
  \item{\strong{watershed} - Name of watershed}
}
}

\section{Instream}{

Datasets containing the instream rearing habitat area in square meters as a
function of flow in cubic feet per second.

\itemize{
  \item \code{upper_sacramento_river_instream}
  \item \code{upper_mid_sacramento_river_instream}
  \item \code{lower_mid_sacramento_river_instream}
  \item \code{lower_sacramento_river_instream}
}

Data Format:
\itemize{
  \item{\strong{flow_cfs} - Flow value in cubic feet per second}
  \item{\strong{rearing_sq_meters} - Suitable rearing area in square meteres}
  \item{\strong{watershed} - Name of watershed}
}
}

\section{Floodplain}{

Datasets containing the floodplain habitat area in square meters as a function
of flow in cubic feet per second.

\itemize{
  \item \code{upper_sacramento_river_floodplain}
  \item \code{upper_mid_sacramento_river_floodplain}
  \item \code{lower_mid_sacramento_river_floodplain}
  \item \code{lower_sacramento_river_floodplain}
}

Data Format:
\itemize{
  \item{\strong{flow_cfs} - Flow value in cubic feet per second}
  \item{\strong{floodplain_sq_meters} - Suitable rearing area in square meteres}
  \item{\strong{watershed} - Name of watershed}
}
}

\keyword{datasets}
