% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-bypass-habitat.R
\name{set_bypass_floodplain_habitat}
\alias{set_bypass_floodplain_habitat}
\title{Set Bypass Floodplain Habitat Area}
\usage{
set_bypass_floodplain_habitat(bypass, flow)
}
\arguments{
\item{bypass}{one section of the bypasses ('yolo1', 'yolo2', 'sutter1', 'sutter2', 'sutter3', or 'sutter4')}

\item{flow}{value used to determine habitat area}
}
\value{
habitat area in square meters
}
\description{
This function returns an estimated total floodplain habitat area based on bypass and flow.
}
\examples{
# habitat area in square meters for section 1 of the yolo bypass at 200 cfs
set_bypass_floodplain_habitat('yolo1', 200000)
}
