% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-instream-habitat.R
\name{set_instream_area}
\alias{set_instream_area}
\title{Set instream habitat area based on watershed, species, life stage and flow}
\usage{
set_instream_area(watershed, species, life_stage, flow)
}
\arguments{
\item{watershed}{one of the watersheds defined for the SIT model}

\item{species}{one of 'fr', 'sr', or 'st'}

\item{life_stage}{life stage of fish, one of 'juv', 'adult' or 'fry'}

\item{flow}{value used to determine habitat area}
}
\description{
Set instream habitat area based on watershed, species, life stage and flow
}
