% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-spawning-habitat.R
\name{set_spawning_area}
\alias{set_spawning_area}
\title{Set spawning habitat area based on watershed, species, life stage and flow.}
\usage{
set_spawning_area(watershed, species, flow)
}
\arguments{
\item{watershed}{one of the watersheds defined for the SIT model}

\item{species}{one of 'fr' (Fall Run), 'sr' (Spring Run), or 'st' (Steelhead)}

\item{flow}{a flow value in cubic feet per second used to determine habitat area}
}
\value{
habitat area in square meters
}
\description{
Set spawning habitat area based on watershed, species, life stage and flow.
}
