% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-spawning-habitat.R, R/set-spawning.R
\name{spawning_approx}
\alias{spawning_approx}
\alias{spawning_approx}
\title{function uses an existing relationship to return a linear interpolated approx function}
\usage{
spawning_approx(watershed, species = "fr")

spawning_approx(watershed, species = "fr")
}
\arguments{
\item{watershed}{name of the watershed to compute approx function on}

\item{species}{one of 'fr' (Fall Run), 'sr' (Spring Run), or 'st' (Steelhead)}

\item{watershed}{name of the watershed to compute approx function on}

\item{species}{one of 'fr' (Fall Run), 'sr' (Spring Run), or 'st' (Steelhead)}
}
\value{
an approx function obtained from calling \code{\link[stats]{approxfun}}

an approx function obtained from calling \code{\link[stats]{approxfun}}
}
\description{
function uses an existing relationship to return a linear interpolated approx function

function uses an existing relationship to return a linear interpolated approx function
}
