% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-instream-habitat.R
\name{ST_rearing_approx}
\alias{ST_rearing_approx}
\title{Steelhead rearing habitat flow to area approximator}
\usage{
ST_rearing_approx(relationship_df, modeling_lookup, life_stage)
}
\arguments{
\item{relationship_df}{dataframe from cvpiaHabitat with a flow to wua relationship}

\item{modeling_lookup}{modeling lookup dataframe from cvpiaHabitat}

\item{life_stage}{one of 'spawn', 'juv' or 'fry'}
}
\description{
function creates the approx function for spring run
}
