% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stony_creek_floodplain}
\alias{stony_creek_floodplain}
\title{Stony Creek Flow to Floodplain Habitat Area Relationship}
\format{dataframe with 32 rows and 4 variables:
\describe{
  \item{flow_cfs}{flow value in cubic feet per second}
  \item{FR_floodplain_acres}{fall run floodplain acreage}
  \item{ST_floodplain_acres}{steelhead floodplain acreage}
  \item{watershed}{name of watershed}
}}
\source{
\href{#}{Scaled from a Cottonwood Creek flow to floodplain area relationship generated with a hybrid USFWS / FEMA 1D HEC-RAS hydraulic model}
}
\usage{
stony_creek_floodplain
}
\description{
A dataset containing total inundated floodplain area in acres as a function of flow in cubic feet per second.
}
\section{Fall Run and Steelhead Modeling}{

No site-specific hydraulic modeling was available for this watershed. A flow to floodplain area relationship was generated for this watershed by scaling the flow to floodplain area relationship for Cottonwood Creek  based on hydrologic and geomorphic analyses. Flows and corresponding floodplain areas per unit length were calculated for this watershed as 0.59 percent of Cottonwood Creek values based on the ratio of mean flow from December to June between this watershed and Cottonwood Creek. Of the entire mapped 26.12 mile rearing extent in this watershed, 26.12 miles were classified as low gradient and 0.00 miles were classified as high gradient based on a geomorphic analysis. A scaling factor of 0.1 was  applied to the area per unit length for the high gradient extent. No scaling factor was applied to the low gradient extent.
}

\examples{
stony_creek_floodplain

}
\keyword{datasets}
