% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sutter_bypass_floodplain}
\alias{sutter_bypass_floodplain}
\title{Sutter Bypass Flow to Floodplain Habitat Area Relationship}
\format{dataframe with 41 rows and 5 variables:
\describe{
  \item{flow_cfs}{integer flow value in cubic feet per second}
  \item{Sutter Bypass 1}{floodplain acreage in the Sutter Bypass, to Moulton Weir}
  \item{Sutter Bypass 2}{floodplain acreage in the Sutter Bypass, to Colusa Weir}
  \item{Sutter Bypass 3}{floodplain acreage in the Sutter Bypass, to Tisdale Weir}
  \item{Sutter Bypass 4}{floodplain acreage in the Sutter Bypass below Tisdale Weir}
}}
\source{
\href{https://s3-us-west-2.amazonaws.com/cvpiahabitat-r-package/cvpia-sit-model-inputs/HendrixEtAl2014_Winter_Run_Model_Tech_Memo.pdf}{Central Valley Floodplain Evaluation and Delineation (CVFED) HEC-RAS hydraulic model refined for use in the NOAA-NMFS Winter Run Chinook Salmon life cycle model}
}
\usage{
sutter_bypass_floodplain
}
\description{
A dataset containing the floodplain habitat area in acres as a
function of flow in cubic feet per second
}
\details{
Habitat estimates from Correigh Greene's Winter Run Life Cycle Model

High quality defined by:

Channel depth > 0.2 m and < 1.5 m

Velocity <= 0.15 m/s
}
\examples{
sutter_bypass_floodplain

}
\keyword{datasets}
