% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tuolumne_river_instream}
\alias{tuolumne_river_instream}
\title{Tuolumne River Instream Flow to Habitat Area Relationship}
\format{dataframe with 30 rows and 4 variables:
\describe{
\item{flow_cfs}{integer flow value in cubic feet per second}
\item{spawn_WUA}{chinook spawning WUA in square feet per 1000 feet}
\item{fry_WUA}{chinook fry (up to 50 mm) WUA in square feet per 1000 feet}
\item{juv_WUA}{chinook juvenile WUA in square feet per 1000 feet}
\item{watershed}{name of watershed}
}}
\source{
\href{https://s3-us-west-2.amazonaws.com/cvpiahabitat-r-package/cvpia-sit-model-inputs/Stillwater2010_Tuolumne_P-2299-075_71_DP_FLA_AttC_StudyRept_IFIM_AppA-H_140428.pdf}{Flo2D model by Stillwater, 2014}
}
\usage{
tuolumne_river_instream
}
\description{
A dataset containing the Weighted Usable Area (WUA) in square feet per 1000 feet
as a function of flow in cubic feet per second
}
\details{
The spawning and rearing WUAs were derived using a PHABSIM model by Stillwater 2010.
}
\keyword{datasets}
