% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{upper_mid_sac_region_instream}
\alias{upper_mid_sac_region_instream}
\title{Upper-mid Sacramento River Region Instream Flow to Habitat Area Relationship}
\format{dataframe with 14 rows and 5 variables:
\describe{
  \item{flow_cfs}{integer flow value in cubic feet per second}
  \item{FR_spawn_wua}{spawning WUA in square feet per 1000 feet}
  \item{FR_fry_wua}{fry (up to 50 mm) WUA in square feet per 1000 feet}
  \item{FR_juv_wua}{juvenile WUA in square feet per 1000 feet}
  \item{watershed}{name of watershed}
}}
\source{
Sadie Gill
}
\usage{
upper_mid_sac_region_instream
}
\description{
A dataset containing the Weighted Usable Area (WUA) in square feet per 1000 feet
as a function of flow in cubic feet per second
}
\details{
Some of the Upper Sacramento Rivers' instream rearing habitat have not been modeled.
The rearing WUA values were estimated using the mean WUA at each flow from Battle Creek,
Butte Creek, Clear Creek, and Cow Creek. The spawning WUA values were estimated using the mean
WUA at each flow from Battle Creek, Butte Creek, and Clear Creek.

Spawning Not Modeled:

\itemize{
  \item Antelope Creek
  \item Bear Creek
  \item Big Chico Creek
  \item Cow Creek
  \item Deer Creek
  \item Elder Creek
  \item Mill Creek
  \item Paynes Creek
  \item Stony Creek
  \item Thomes Creek
}

Rearing Not Modeled:

\itemize{
  \item Antelope Creek
  \item Bear Creek
  \item Big Chico Creek
  \item Deer Creek
  \item Elder Creek
  \item Mill Creek
  \item Paynes Creek
  \item Stony Creek
  \item Thomes Creek
}
}
\examples{
upper_mid_sac_region_instream

}
\keyword{datasets}
