% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{upper_mid_sacramento_river_floodplain}
\alias{upper_mid_sacramento_river_floodplain}
\title{Upper-mid Sacramento River Flow to Floodplain Habitat Area Relationship}
\format{dataframe with 240 rows and 3 variables:
\describe{
  \item{flow_cfs}{integer flow value in cubic feet per second}
  \item{FR_floodplain_acres}{fall run floodplain acreage}
  \item{watershed}{name of watershed}
}}
\source{
\href{https://www.google.com}{NMFS Life Cycle Model, Hendrix 2017 [missing url]}
}
\usage{
upper_mid_sacramento_river_floodplain
}
\description{
A dataset containing the floodplain habitat area in acres as a
function of flow in cubic feet per second
}
\details{
The habitat area was derived from HEC-RAS 1D model, Hendrix 2017.
The study's extent is from Battle Creek to the confluence with the Feather River (189.1 mi).
The CVPIA Upper-mid Sacramento River extends from Red Blurr to Wilkins Slough (122.45 mi).
The study results are scaled to the CVPIA extent.
}
\keyword{datasets}
