% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{upper_mid_sacramento_river_instream}
\alias{upper_mid_sacramento_river_instream}
\title{Upper-Mid Sacramento River Instream Flow to Habitat Area Relationship}
\format{dataframe with 45 rows and 3 variables:
\describe{
  \item{flow_cfs}{integer flow value in cubic feet per second}
  \item{FR_juv_wua}{juvenile WUA in square feet per 1000 feet}
  \item{watershed}{name of watershed}
}}
\source{
NMFS Life Cycle Model, Hendrix 2017
}
\usage{
upper_mid_sacramento_river_instream
}
\description{
A dataset containing the Weighted Usable Area (WUA) in square feet per 1000 feet
as a function of flow in cubic feet per second
}
\details{
The rearing WUA was derived from a HEC-RAS 1D model. The study's extent is from Battle Creek to
the confluence with the Feather River (189.1 mi). The CVPIA Upper-mid Sacramento River extends
from Red Blurr to Wilkins Slough (122.45 mi).

\strong{NOTE:} The CVPIA Lower-mid Sacramento River segment is from Wilkins Slough to the
American River (58.0 mi). To calculate the available habitat, use a proportional combination of
the WUA from the Upper-mid (38.3 mi, 2/3) and Lower (19.7 mi, 1/3) Sacramento instream values.
}
\keyword{datasets}
