% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{upper_sacramento_river_floodplain}
\alias{upper_sacramento_river_floodplain}
\title{Upper Sacramento River Flow to Floodplain Habitat Area Relationship}
\format{dataframe with 240 rows and 3 variables:
\describe{
  \item{flow_cfs}{integer flow value in cubic feet per second}
  \item{FR_floodplain_acres}{fall run floodplain acreage}
  \item{watershed}{name of watershed}
}}
\source{
\href{https://s3-us-west-2.amazonaws.com/cvpiahabitat-r-package/cvpia-sit-model-inputs/HendrixEtAl2014_Winter_Run_Model_Tech_Memo.pdf}{Central Valley Floodplain Evaluation and Delineation (CVFED) HEC-RAS hydraulic model refined for use in the NOAA-NMFS Winter Run Chinook Salmon life cycle model}
}
\usage{
upper_sacramento_river_floodplain
}
\description{
A dataset containing the floodplain habitat area in acres as a
function of flow in cubic feet per second
}
\details{
The habitat area was derived from a HEC-RAS 1D model, Hendrix 2017.
The CVPIA Upper Sacramento River extends from Keswick to Red Bluff (59.2 mi).
This reach overlaps with two of the study's, Keswick to Battle Creek (55.5 mi) and Battle Creek to
the confluence with the Feather River (189.1 mi). To scale the study's results to CVPIA's extents,
the floodplain acerage at a given flow is the weighted average of the floodplain
area per river mile within each study.
}
\examples{
upper_sacramento_river_floodplain

}
\keyword{datasets}
