% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{watershed_lengths}
\alias{watershed_lengths}
\title{Habitat Extent Lengths}
\format{dataframe with 57 rows and 7 variables
\describe{
\item{order}{integer value representing watershed order in SIT model}
\item{watershed}{name of watershed}
\item{lifestage}{habitat type by lifestage, 'spawing' or 'rearing'}
\item{miles}{length in miles}
\item{feet}{length in feet}
\item{source}{Expert who delineated habitat extents}
\item{species}{species of habitat extent}
}}
\source{
FlowWest, add url to shp file when mapping is complete
}
\usage{
watershed_lengths
}
\description{
A dataset containing the length of rearing and spawning extent within each CVPIA watershed for different species.
}
\details{
Information comes from expert outreach undertaken in 2017.
}
\examples{
watershed_lengths

}
\keyword{datasets}
