% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{weeks_flooded}
\alias{weeks_flooded}
\title{Duration of Floodplain Inundation}
\usage{
weeks_flooded(ws, flow_cfs)
}
\arguments{
\item{ws}{name of CVPIA watershed}

\item{flow_cfs}{monthly mean flow in cubic feet per second}
}
\value{
integer value for number of weeks
}
\description{
Function returns number of weeks of floodplain inundation given a flow for a watershed
}
\details{
These relationships between number of days inundated and the mean monthly flow
are stored in \code{\link{weeks_inundated}} and referenced by this function.
For watersheds without a defined relationship, a two week inundation duration is assumed.
}
\examples{
weeks_flooded('Yuba River', 900)

}
