% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wua_to_area}
\alias{wua_to_area}
\title{WUA to Area in Square Meters}
\usage{
wua_to_area(wua, watershed_name, life_stage)
}
\arguments{
\item{wua}{WUA in square feet per 1000 feet}

\item{life_stage}{life stage}

\item{watershed}{watershed}

\item{species}{species}
}
\description{
Function converts a Weighted Usable Area (WUA) to an
area in square meters by multiplying the WUA by the stream's length used by the species.
The lengths were obtained through expert outreach in 2017 and are available here \code{\link{cvpiaHabitat::watershed_lengths}}
}
