% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yolo_bypass_instream}
\alias{yolo_bypass_instream}
\title{Yolo Bypass Instream Flow to Habitat Area Relationship}
\format{dataframe with 41 rows and 3 variables:
\describe{
  \item{flow_cfs}{integer flow value in cubic feet per second}
  \item{Yolo Bypass 1}{juvenile rearing habitat WUA in the Yolo Bypass, Fremont Weir to Sacramento Weir}
  \item{Yolo Bypass 2}{juvenile rearing habitat WUA in the Yolo Bypass below Sacramento Weir}
}}
\source{
\href{https://s3-us-west-2.amazonaws.com/cvpiahabitat-r-package/cvpia-sit-model-inputs/HendrixEtAl2014_Winter_Run_Model_Tech_Memo.pdf}{Central Valley Floodplain Evaluation and Delineation (CVFED) HEC-RAS hydraulic model refined for use in the NOAA-NMFS Winter Run Chinook Salmon life cycle model}
}
\usage{
yolo_bypass_instream
}
\description{
A dataset containing the Weighted Usable Area (WUA) in square feet
as a function of flow in cubic feet per second
}
\details{
Habitat estimates from Correigh Greene's Winter Run Life Cycle Model

High quality defined by:

Channel depth > 0.2 m and < 1.5 m

Velocity <= 0.15 m/s
}
\examples{
yolo_bypass_instream

}
\keyword{datasets}
