% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yuba_river_instream}
\alias{yuba_river_instream}
\title{Yuba River Instream Flow to Habitat Area Relationship}
\format{dataframe with 34 rows and 10 variables
\describe{
\item{flow_cfs}{numeric flow value in cubic feet per second}
\item{FR_spawning}{Fall Run spawning WUA in square feet per 1000 feet}
\item{FR_SR_fry}{Fall and Spring Run fry WUA in square feet per 1000 feet}
\item{FR_SR_juv}{Fall and Spring Run juvenile WUA in square feet per 1000 feet}
\item{SR_spawning}{Spring Run spawning WUA in square feet per 1000 feet}
\item{ST_fry}{Steelhead fry WUA in square feet per 1000 feet}
\item{ST_juv}{Steelhead juvenile WUA in square feet per 1000 feet}
\item{ST_spawning}{Steelhead spawning WUA in square feet per 1000 feet}
\item{watershed}{name of watershed}
}}
\usage{
yuba_river_instream
}
\description{
A dataset containing the Weighted Usable Area (WUA) in square feet per 1000 feet
as a function of flow in cubic feet per second
}
\keyword{datasets}
