% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{american_river_instream}
\alias{american_river_instream}
\title{American River Instream Flow to Habitat Area Relationship}
\format{dataframe with 11 rows and 5 variables:
\describe{
\item{flow_cfs}{integer flow value in cubic feet per second}
\item{FR_spawn_wua}{chinook spawning WUA in square feet per 1000 feet}
\item{FR_fry_wua}{chinook fry (up to 50 mm) WUA in square feet per 1000 feet}
\item{FR_juv_wua}{chinook juvenile WUA in square feet per 1000 feet}
\item{watershed}{name of watershed}
}}
\source{
\href{TODO}{FWS 1985}
}
\usage{
american_river_instream
}
\description{
A dataset containing the Weighted Usable Area (WUA) in square feet per 1000 feet
as a function of flow in cubic feet per second
}
\details{
The spawning WUA was derived from a River2D,
applied habitat suitability model.
}
\examples{
american_river_instream

}
\keyword{datasets}
