% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bear_river_floodplain}
\alias{bear_river_floodplain}
\title{Bear River Flow to Floodplain Habitat Area Relationship}
\format{dataframe with 37 rows and 4 variables:
\describe{
  \item{flow_cfs}{integer flow value in cubic feet per second}
  \item{FR_floodplain_acres}{fall run floodplain acerage}
  \item{ST_floodplain_acres}{steelhead floodplain acerage}
  \item{watershed}{name of watershed}
}}
\source{
NA
}
\usage{
bear_river_floodplain
}
\description{
A dataset containing the floodplain habitat area in acres as a
function of flow in cubic feet per second
}
\details{
NA
}
\keyword{datasets}
