% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{big_chico_creek_floodplain}
\alias{big_chico_creek_floodplain}
\title{Big Chico Creek Flow to Floodplain Habitat Area Relationship}
\format{dataframe with 32 rows and 3 variables:
\describe{
  \item{flow_cfs}{integer flow value in cubic feet per second}
  \item{FR_floodplain_acres}{fall run floodplain acreage (use for Steelhead when needed)}
  \item{watershed}{name of watershed}
}}
\source{
\href{https://s3-us-west-2.amazonaws.com/cvpiahabitat-r-package/cvpia-sit-model-inputs/CVFPP2012_Att8_June.pdf}{CVFPP 2012}
}
\usage{
big_chico_creek_floodplain
}
\description{
A dataset containing the floodplain habitat area in acres as a
function of flow in cubic feet per second
}
\details{
The habitat area was derived from the CVFPP 2012 HEC-RAS 1D model.
}
\keyword{datasets}
