% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{butte_creek_floodplain}
\alias{butte_creek_floodplain}
\title{Butte Creek Flow to Floodplain Habitat Area Relationship}
\format{dataframe with 31 rows and 4 variables:
\describe{
  \item{flow_cfs}{integer flow value in cubic feet per second}
  \item{FR_floodplain_acres}{fall run floodplain acerage}
  \item{SR_floodplain_acres}{spring run floodplain acerage}
  \item{ST_floodplain_acres}{steelhead floodplain acerage}
  \item{watershed}{name of watershed}
}}
\source{
CVFPP 2012 \url{https://s3-us-west-2.amazonaws.com/cvpiahabitat-r-package/cvpia-sit-model-inputs/CVFPP2012_Att8_June.pdf}
}
\usage{
butte_creek_floodplain
}
\description{
A dataset containing the floodplain habitat area in acres as a
function of flow in cubic feet per second
}
\details{
The habitat area was derived from the CVFPP 2012 HEC-RAS 1D model.
The total area * suitability (25%)
}
\keyword{datasets}
