% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{butte_creek_instream}
\alias{butte_creek_instream}
\title{Butte Creek Instream Flow to Habitat Area Relationship}
\format{dataframe with 41 rows and 6 variables:
\describe{
  \item{flow_cfs}{integer flow value in cubic feet per second}
  \item{spawn_WUA}{spawning WUA in square feet per 1000 feet}
  \item{fry_WUA}{fry (up to 50 mm) WUA in square feet per 1000 feet}
  \item{juv_WUA}{juvenile WUA in square feet per 1000 feet}
  \item{adult_trout_WUA}{adult trout WUA in square feet per 1000 feet}
  \item{watershed}{name of watershed}
}}
\source{
FWS and 2008 FERC relicensing of DeSabla
}
\usage{
butte_creek_instream
}
\description{
A dataset containing the Weighted Usable Area (WUA) in square feet per 1000 feet
as a function of flow in cubic feet per second
}
\details{
The spawning WUA was derived from a FWS River2D model and the rearing WUA from a PHABSIM
model created for the 2008 FERC relicensing of DeSabla. The spawning values are the summation of two river
segments, above Centerville Powerhouse (6.5 miles) and below (9 miles).
The spawning results are for Spring Run Chinook. The rearing results are for Fall Run Chinook.
}
\keyword{datasets}
