% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clear_creek_instream}
\alias{clear_creek_instream}
\title{Clear Creek Instream Flow to Habitat Area Relationship}
\format{
dataframe with 23 rows and 11 variables:
\describe{
  \item{flow_cfs}{integer flow value in cubic feet per second}
  \item{FR_fry_wua}{fall run fry (up to 50 mm) WUA in square feet per 1000 feet}
  \item{FR_juv_wua}{fall run juvenile WUA in square feet per 1000 feet}
  \item{FR_spawn_wua}{fall run spawning WUA in square feet per 1000 feet}
  \item{SR_fry_wua}{spring run fry (up to 50 mm) WUA in square feet per 1000 feet}
  \item{SR_juv_wua}{spring run fry juvenile WUA in square feet per 1000 feet}
  \item{SR_spawn_wua}{spring run fry spawning WUA in square feet per 1000 feet}
  \item{ST_fry_wua}{steelhead fry (up to 50 mm) WUA in square feet per 1000 feet}
  \item{ST_juv_wua}{steelhead juvenile WUA in square feet per 1000 feet}
  \item{ST_spawn_wua}{steelhead spawning WUA in square feet per 1000 feet}
  \item{watershed}{name of watershed}
}
}
\source{
\href{https://s3-us-west-2.amazonaws.com/cvpiahabitat-r-package/cvpia-sit-model-inputs/FWS2007-2013_ClearCreekInstream.pdf}{FWS 2007-2013}
}
\usage{
clear_creek_instream
}
\description{
A dataset containing the Weighted Usable Area (WUA) in square feet per 1000 feet
as a function of flow in cubic feet per second
}
\details{
The spawning and rearing WUA were derived from a FWS River2D model. The WUA values are from
three river segments, Upper Alluvial (2.27 miles), Canyon Segment (7.33 miles), and Lower Alluvial (8.81 miles).
Fall Run are only in the Lower Alluvial Segment and there is a segregation weir that prevents the
Spring Run from spawning in the Lower Alluvial Segment.
}
\examples{
clear_creek_instream

}
\keyword{datasets}
