% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cosumnes_river_floodplain}
\alias{cosumnes_river_floodplain}
\title{Cosumnes River Flow to Floodplain Habitat Area Relationship}
\format{dataframe with 36 rows and 4 variables:
\describe{
  \item{flow_cfs}{integer flow value in cubic feet per second}
  \item{FR_floodplain_acres}{fall run floodplain acreage}
  \item{ST_floodplain_acres}{steelhead floodplain acreage}
  \item{watershed}{name of watershed}
}}
\source{
NA
}
\usage{
cosumnes_river_floodplain
}
\description{
A dataset containing the floodplain habitat area in acres as a
function of flow in cubic feet per second
}
\details{
NA
}
\keyword{datasets}
