% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deer_creek_floodplain}
\alias{deer_creek_floodplain}
\title{Deer Creek Flow to Floodplain Habitat Area Relationship}
\format{
dataframe with 30 rows and 5 variables:
\describe{
  \item{flow_cfs}{integer flow value in cubic feet per second}
  \item{FR_floodplain_acres}{fall run floodplain acreage}
  \item{SR_floodplain_acres}{spring run floodplain acreage}
  \item{ST_floodplain_acres}{steelhead floodplain acreage}
  \item{watershed}{name of watershed}
}
}
\source{
\href{https://s3-us-west-2.amazonaws.com/cvpiahabitat-r-package/cvpia-sit-model-inputs/DRAFT+Deer+Creek+Hydraulic+Models+Tech+Memo+6-08-07.pdf}{Deer Creek Watershed Conservancy (DCWC) 1D HEC-RAS model}
}
\usage{
deer_creek_floodplain
}
\description{
A dataset containing the floodplain habitat area in acres as a
function of flow in cubic feet per second
}
\section{Fall Run Modeling}{

A 11.76 mile portion of the entire mapped rearing extent of 18.39 miles was modeled using Deer Creek Watershed Conservancy (DCWC) 1D HEC-RAS model. Of the entire mapped rearing extent, 11.76 miles were classified as low gradient and 6.63 miles were classified as high gradient based on a geomorphic analysis. Floodplain area per unit length was determined for the modeled extent and applied to determine areas for the non-modeled extent. A scaling factor of 0.1 was applied to the area per unit length for the high gradient extent. No scaling factor was applied to the low gradient extent.
}

\section{Spring Run and Steelhead Modeling}{

A 11.76 mile portion of the entire mapped rearing extent of 47.92 miles was modeled using Deer Creek Watershed Conservancy (DCWC) 1D HEC-RAS model. Of the entire mapped rearing extent, 11.76 miles were classified as low gradient and 36.16 miles were classified as high gradient based on a geomorphic analysis. Floodplain area per unit length was determined for the modeled extent and applied to determine areas for the non-modeled extent. A scaling factor of 0.1 was applied to the area per unit length for the high gradient extent. No scaling factor was applied to the low gradient extent.
}

\examples{
deer_creek_floodplain
}
\keyword{datasets}
