% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{delta_habitat}
\alias{delta_habitat}
\title{Delta Habitat Area}
\format{dataframe with 372 rows and 3 variables
\describe{
  \item{date}{modeled results for 1980-2010}
  \item{North Delta}{high quality habitat area in square meters}
  \item{South Delta}{high quality habitat area in square meters}
}}
\source{
\href{https://s3-us-west-2.amazonaws.com/cvpiahabitat-r-package/cvpia-sit-model-inputs/HendrixEtAl2014_Winter_Run_Model_Tech_Memo.pdf}{Central Valley Floodplain Evaluation and Delineation (CVFED) HEC-RAS hydraulic model refined for use in the NOAA-NMFS Winter Run Chinook Salmon life cycle model}

\itemize{
  \item Modeling Output: Correigh Green \email{correigh.greene@noaa.gov}
  \item Delta Node Selection: Mark Tompkins \email{mtompkins@flowwest.com}
  \item Data Wrangling: Sadie Gill \email{sgill@flowwest.com}
}
}
\usage{
delta_habitat
}
\description{
A dataset containing the area of highly suitable habitat within the
North and South Deltas
}
\details{
Habitat estimates from NOAA NMFS Winter Run Life Cycle Model. The model
outputs habitat area estimates for months December-May. Habitat for months
June-Novemeber are represented by the yearly average habitat.

High quality habitat defined by:

Channel type: mainstem, distributaries, or open water

Depth: > 0.2 m and <= 1.5 m

Cover Vegetated edge
}
\keyword{datasets}
