% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-instream-habitat.R
\name{FR_rearing_approx}
\alias{FR_rearing_approx}
\title{function creates the approx function for fall run}
\usage{
FR_rearing_approx(relationship_df, modeling_lookup, life_stage)
}
\arguments{
\item{relationship_df}{dataframe from cvpiaHabitat with a flow to wua relationship}

\item{modeling_lookup}{modeling lookup dataframe from cvpiaHabitat}
}
\description{
function creates the approx function for fall run
}
