% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-spawning-habitat.R
\name{FR_spawn_approx}
\alias{FR_spawn_approx}
\title{function creates the approx function for fall run}
\usage{
FR_spawn_approx(relationship_df, modeling_lookup)
}
\arguments{
\item{relationship_df}{dataframe from cvpiaHabitat with a flow to wua relationship}

\item{modeling_lookup}{modeling lookup dataframe from cvpiaHabitat}
}
\description{
function creates the approx function for fall run
}
