% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lower_sacramento_river_floodplain}
\alias{lower_sacramento_river_floodplain}
\title{Lower Sacramento River Flow to Floodplain Habitat Area Relationship}
\format{dataframe with 240 rows and 3 variables:
\describe{
  \item{flow_cfs}{integer flow value in cubic feet per second}
  \item{FR_floodplain_acres}{fall run floodplain acreage}
  \item{watershed}{name of watershed}
}}
\source{
\href{https://www.google.com}{missing}
}
\usage{
lower_sacramento_river_floodplain
}
\description{
A dataset containing the floodplain habitat area in acres as a
function of flow in cubic feet per second
}
\details{
The habitat area was derived from a HEC-RAS 1D model. The study's extent is
from the confluence with the Feather River to Freeport (33.4 mi). The CVPIA Lower
Sacramento River extends from the confluence with the American River to Freeport (13.7 mi).
The study results are scaled to the CVPIA extent.
}
\examples{
lower_sacramento_river_floodplain

}
\keyword{datasets}
