% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{merced_river_floodplain}
\alias{merced_river_floodplain}
\title{Merced River Flow to Floodplain Habitat Area Relationship}
\format{
dataframe with 22 rows and 3 variables:
\describe{
  \item{flow_cfs}{integer flow value in cubic feet per second}
  \item{floodplain_acres}{fall run floodplain acreage (use for Steelhead when required)}
  \item{watershed}{name of watershed} }
}
\source{
\href{https://s3-us-west-2.amazonaws.com/cvpiahabitat-r-package/cvpia-sit-model-inputs/TO25-ST8_Tech_Memo_SJR.pdf}{Central Valley Floodplain Evaluation and Delineation (CVFED) HEC-RAS hydraulic model}
}
\usage{
merced_river_floodplain
}
\description{
A dataset containing the floodplain habitat area in acres as a
  function of flow in cubic feet per second
}
\section{Fall Run and Steelhead Modeling}{

A 25.50 mile portion of the entire mapped rearing extent of 51.76 miles was modeled using Central Valley Floodplain Evaluation and Delineation (CVFED) HEC-RAS hydraulic model. Of the entire mapped rearing extent, 34.60 miles were classified as low gradient and 17.16 miles were classified as high gradient based on a geomorphic analysis. Floodplain area per unit length was determined for the modeled extent and applied to determine areas for the non-modeled extent. A scaling factor of 0.1 was applied to the area per unit length for the high gradient extent. No scaling factor was applied to the low gradient extent.
}

\examples{
merced_river_floodplain
}
\keyword{datasets}
