% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mokelumne_river_floodplain}
\alias{mokelumne_river_floodplain}
\title{Mokelumne River to Floodplain Habitat Area Relationship}
\format{dataframe with 31 rows and 5 variables:
\describe{
  \item{flow_cfs}{integer flow value in cubic feet per second}
  \item{FR_floodplain_acres}{fall run floodplain acreage}
  \item{SR_floodplain_acres}{spring run floodplain acreage}
  \item{ST_floodplain_acres}{steelhead floodplain acreage}
  \item{watershed}{name of watershed}
}}
\source{
NA
}
\usage{
mokelumne_river_floodplain
}
\description{
A dataset containing the floodplain habitat area in acres as a
function of flow in cubic feet per second
}
\details{
NA
}
\keyword{datasets}
