% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mokelumne_river_floodplain}
\alias{mokelumne_river_floodplain}
\title{Mokelumne River to Floodplain Habitat Area Relationship}
\format{dataframe with 31 rows and 3 variables:
\describe{
  \item{flow_cfs}{integer flow value in cubic feet per second}
  \item{FR_floodplain_acres}{fall run floodplain acreage (use for Spring Run and Steelhead when required)}
  \item{watershed}{name of watershed}
}}
\source{
NA
}
\usage{
mokelumne_river_floodplain
}
\description{
A dataset containing the floodplain habitat area in acres as a
function of flow in cubic feet per second
}
\details{
NA
}
\examples{
mokelumne_river_floodplain
}
\keyword{datasets}
