% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mokelumne_river_instream}
\alias{mokelumne_river_instream}
\title{Mokelumne River Instream Flow to Habitat Area Relationship}
\format{dataframe with 14 rows and 5 variables:
\describe{
  \item{flow_cfs}{integer flow value in cubic feet per second}
  \item{FR_spawn_wua}{spawning WUA in square feet per 1000 feet}
  \item{FR_fry_wua}{fry (up to 50 mm) WUA in square feet per 1000 feet}
  \item{FR_juv_wua}{juvenile WUA in square feet per 1000 feet}
  \item{watershed}{name of watershed}
}}
\source{
EBMUD 2016 and CDFW 1998
}
\usage{
mokelumne_river_instream
}
\description{
A dataset containing the Weighted Usable Area (WUA) in square feet per 1000 feet
as a function of flow in cubic feet per second
}
\details{
The spawning WUA values were derived from the EBMUD 2016 SRH2D model and the rearing from
the CDFW 1998 PHABSIM.
}
\examples{
mokelumne_river_instream

}
\keyword{datasets}
