% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-floodplain-habitat.R
\name{set_floodplain_habitat}
\alias{set_floodplain_habitat}
\title{Set Floodplain Habitat Area}
\usage{
set_floodplain_habitat(watershed, species, flow)
}
\arguments{
\item{watershed}{a watershed defined for the SIT model}

\item{species}{one of 'fr' (Fall Run), 'sr' (Spring Run), or 'st' (Steelhead)}

\item{flow}{a flow value in cubic feet per second}
}
\value{
floodplain habitat value in square meters
}
\description{
This function returns an estimated floodplain area based on watershed, species and flow.
}
\examples{
# floodplain habitat value in square meters for Fall Run in the American River
set_floodplain_habitat("American River", "fr", 34652)

# floodplain habitat value in square meters for Steelhead in the American River
set_floodplain_habitat("American River", "st", 34652)
}
