% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stanislaus_river_instream}
\alias{stanislaus_river_instream}
\title{Stanislaus River Instream Flow to Habitat Area Relationship}
\format{dataframe with 47 rows and 7 variables:
\describe{
  \item{flow_cfs}{integer flow value in cubic feet per second}
  \item{FR_spawn_wua}{spawning WUA in square feet per 1000 feet}
  \item{FR_fry_wua}{fall run fry (up to 50 mm) WUA in square feet per 1000 feet}
  \item{FR_juv_wua}{fall run juvenile WUA in square feet per 1000 feet}
  \item{ST_fry_wua}{steelhead fry (up to 50 mm) WUA in square feet per 1000 feet}
  \item{ST_juv_wua}{steelhead juvenile WUA in square feet per 1000 feet}
  \item{watershed}{name of watershed}
}}
\source{
FWS
}
\usage{
stanislaus_river_instream
}
\description{
A dataset containing the Weighted Usable Area (WUA) in square feet per 1000 feet
as a function of flow in cubic feet per second
}
\details{
The spawning WUA was derived from the Aceituno 1993 (FWS) PHABSIM model and the rearing from
the FWS River2D model.
}
\examples{
stanislaus_river_instream

}
\keyword{datasets}
