% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tuolumne_river_floodplain}
\alias{tuolumne_river_floodplain}
\title{Tuolumne River Flow to Floodplain Habitat Area Relationship}
\format{dataframe with 37 rows and 3 variables:
\describe{
  \item{flow_cfs}{integer flow value in cubic feet per second}
  \item{floodplain_acres}{fall run floodplain acreage (use for Spring Run and Steelhead when required)}
  \item{watershed}{name of watershed}
}}
\source{
TODO fix source
}
\usage{
tuolumne_river_floodplain
}
\description{
A dataset containing the floodplain habitat area in acres as a
function of flow in cubic feet per second
}
\details{
The habitat area was derived from the Flo2D model by Stillwater, 2014.
}
\examples{
tuolumne_river_floodplain

}
\keyword{datasets}
