% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tuolumne_river_floodplain}
\alias{tuolumne_river_floodplain}
\title{Tuolumne River Flow to Floodplain Habitat Area Relationship}
\format{dataframe with 37 rows and 5 variables:
\describe{
  \item{flow_cfs}{integer flow value in cubic feet per second}
  \item{FR_floodplain_acres}{fall run floodplain acerage}
  \item{SR_floodplain_acres}{spring run floodplain acerage}
  \item{ST_floodplain_acres}{steelhead floodplain acerage}
  \item{watershed}{name of watershed}
}}
\source{
\url{https://s3-us-west-2.amazonaws.com/cvpiahabitat-r-package/cvpia-sit-model-inputs/Stillwater2010_Tuolumne_P-2299-075_71_DP_FLA_AttC_StudyRept_IFIM_AppA-H_140428.pdf}
}
\usage{
tuolumne_river_floodplain
}
\description{
A dataset containing the floodplain habitat area in acres as a
function of flow in cubic feet per second
}
\details{
The habitat area was derived from the Flo2D model by Stillwater, 2014.
}
\keyword{datasets}
