% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{upper_mid_sacramento_river_floodplain}
\alias{upper_mid_sacramento_river_floodplain}
\title{Upper-mid Sacramento River Flow to Floodplain Habitat Area Relationship}
\format{dataframe with 240 rows and 3 variables:
\describe{
  \item{flow_cfs}{integer flow value in cubic feet per second}
  \item{FR_floodplain_acres}{fall run floodplain acreage}
  \item{SR_floodplain_acres}{spring run floodplain acreage}
  \item{ST_floodplain_acres}{steelhead floodplain acreage}
  \item{watershed}{name of watershed}
}}
\source{
\href{https://s3-us-west-2.amazonaws.com/cvpiahabitat-r-package/cvpia-sit-model-inputs/HendrixEtAl2014_Winter_Run_Model_Tech_Memo.pdf}{Central Valley Floodplain Evaluation and Delineation (CVFED) HEC-RAS hydraulic model refined for use in the NOAA-NMFS Winter Run Chinook Salmon life cycle model}
}
\usage{
upper_mid_sacramento_river_floodplain
}
\description{
A dataset containing the floodplain habitat area in acres as a
function of flow in cubic feet per second
}
\details{
The habitat area was derived from HEC-RAS 1D model, Hendrix 2017.
The study's extent is from Battle Creek to the confluence with the Feather River (189.1 mi).
The CVPIA Upper-mid Sacramento River extends from Red Blurr to Wilkins Slough (122.45 mi).
The study results are scaled to the CVPIA extent.
}
\section{Fall Run and Steelhead Modeling}{

The entire mapped rearing extent of 59.19 miles was modeled using Central Valley Floodplain Evaluation and Delineation (CVFED) HEC-RAS hydraulic model refined for use in the NOAA-NMFS Winter Run Chinook Salmon life cycle model. The high quality depth and high quality velocity ("Pref11") "BankArea" result was used as floodplain area. High quality velocities were assumed to be less than or equal to 0.15 meters per second, and high quality depths were assumed to be between 0.2 meters and 1.5 meters.
}

\examples{
upper_mid_sacramento_river_floodplain

}
\keyword{datasets}
