% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wua_to_area}
\alias{wua_to_area}
\title{Function converts a weighted usable area (WUA) to area in square meters}
\usage{
wua_to_area(wua, watershed_name, life_stage)
}
\arguments{
\item{wua}{weighted usable area}

\item{life_stage}{life stage}

\item{watershed}{watershed}

\item{species}{species}
}
\description{
Function converts a weighted usable area (WUA) to area in square meters
}
