% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wua_to_area}
\alias{wua_to_area}
\title{Function converts a weighted usable area (WUA) to area in square meters}
\usage{
wua_to_area(wua, ws, sp, ls)
}
\arguments{
\item{wua}{weighted usable area}

\item{ws}{watershed}

\item{sp}{species}

\item{ls}{life stage}
}
\description{
Function converts a weighted usable area (WUA) to area in square meters
}
