% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yolo_bypass_floodplain}
\alias{yolo_bypass_floodplain}
\title{Yolo Bypass Flow to Floodplain Habitat Area Relationship}
\format{dataframe with 2 rows and 3 variables:
\describe{
  \item{flow_cfs}{integer flow value in cubic feet per second}
  \item{Yolo Bypass 1}{floodplain acreage in the Yolo Bypass, Fremont Weir to Sacramento Weir}
  \item{Yolo Bypass 2}{floodplain acreage in the Yolo Bypass below Sacramento Weir}
}}
\source{
\href{https://www.google.com}{TODO}
}
\usage{
yolo_bypass_floodplain
}
\description{
A dataset containing the floodplain habitat area in acres as a
function of flow in cubic feet per second
}
\details{
Habitat estimates from Correigh Greene's Winter Run Life Cycle Model

High quality defined by:

Channel depth > 0.2 m and < 1.5 m

Velocity <= 0.15 m/s
}
\examples{
yolo_bypass_floodplain

}
\keyword{datasets}
