% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yuba_river_floodplain}
\alias{yuba_river_floodplain}
\title{Upper Sacramento River Flow to Floodplain Habitat Area Relationship}
\format{dataframe with 75 rows
Yuba River Flow to Floodplain Habitat Area Relationship}
\source{
\href{https://s3-us-west-2.amazonaws.com/cvpiahabitat-r-package/cvpia-sit-model-inputs/Pasternack2012_LYR+Landforms+Report+(5-9-2012).pdf}{SRH2D}
}
\usage{
yuba_river_floodplain
}
\description{
A dataset containing the floodplain habitat area in acres as a
function of flow in cubic feet per second

A dataset containing the floodplain habitat area in acres as a
function of flow in cubic feet per second
}
\details{
The habitat area was derived from the SRH2D model, Pasternack 2012.
}
\keyword{datasets}
