% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yuba_river_floodplain}
\alias{yuba_river_floodplain}
\title{Yuba River Flow to Floodplain Habitat Area Relationship}
\format{dataframe with 28 rows and 5 variables:
\describe{
  \item{flow_cfs}{integer flow value in cubic feet per second}
  \item{FR_floodplain_acres}{fall run floodplain acerage}
  \item{SR_floodplain_acres}{spring run floodplain acerage}
  \item{ST_floodplain_acres}{steelhead floodplain acerage}
  \item{watershed}{name of watershed}
}}
\source{
SRH2D \url{https://s3-us-west-2.amazonaws.com/cvpiahabitat-r-package/cvpia-sit-model-inputs/Pasternack2012_LYR+Landforms+Report+(5-9-2012).pdf}
}
\usage{
yuba_river_floodplain
}
\description{
A dataset containing the floodplain habitat area in acres as a
function of flow in cubic feet per second
}
\details{
The habitat area was derived from the SRH2D model, Pasternack 2012.
}
\keyword{datasets}
